Handle non-photos post + /start command
Scraper will check if URLs are populated. It will deem as a non-photo
post since logic does not provide any image URLs to the URL string
slice. Since length of the slice returned is 0, it would mean that there
are no photos available, and thus throw an error upwards to print to the
user.